requirements.txt
selenium>=4.15.0
pytest>=7.4.0
allure-pytest>=2.13.0
faker>=21.0.0
